+++ src/main/java/com/stepaniuk/testhorizon/test/TestService.java
@@ -6,12 +6,17 @@ import com.stepaniuk.testhorizon.event.test.TestUpdatedEvent;
 import com.stepaniuk.testhorizon.payload.test.TestCreateRequest;
 import com.stepaniuk.testhorizon.payload.test.TestResponse;
 import com.stepaniuk.testhorizon.payload.test.TestUpdateRequest;
+import com.stepaniuk.testhorizon.project.ProjectRepository;
+import com.stepaniuk.testhorizon.project.exception.NoSuchProjectByIdException;
+import com.stepaniuk.testhorizon.security.authinfo.AuthInfo;
+import com.stepaniuk.testhorizon.shared.exception.AccessToManageEntityDeniedException;
 import com.stepaniuk.testhorizon.shared.PageMapper;
 import com.stepaniuk.testhorizon.test.exceptions.NoSuchTestByIdException;
 import com.stepaniuk.testhorizon.test.exceptions.NoSuchTestTypeByNameException;
 import com.stepaniuk.testhorizon.test.type.TestType;
 import com.stepaniuk.testhorizon.test.type.TestTypeName;
 import com.stepaniuk.testhorizon.test.type.TestTypeRepository;
+import com.stepaniuk.testhorizon.user.authority.AuthorityName;
 import jakarta.annotation.Nullable;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Pageable;
@@ -23,18 +28,28 @@ import java.net.URI;
 import java.time.Instant;
 import java.util.UUID;
 
+import static com.stepaniuk.testhorizon.security.SecurityUtils.hasAuthority;
+import static com.stepaniuk.testhorizon.security.SecurityUtils.isOwner;
+
 @Service
 @RequiredArgsConstructor
 public class TestService {
 
     private final TestRepository testRepository;
     private final TestTypeRepository testTypeRepository;
+    private final ProjectRepository projectRepository;
     private final TestMapper testMapper;
     private final PageMapper pageMapper;
     private final TestProducer testProducer;
 
     public TestResponse createTest(TestCreateRequest testCreateRequest, Long authorId, String correlationId) {
         Test test = new Test();
+
+        var projectId = testCreateRequest.getProjectId();
+
+        if (!projectRepository.existsById(projectId)) {
+            throw new NoSuchProjectByIdException(projectId);
+        }
 
         test.setProjectId(testCreateRequest.getProjectId());
         test.setTestCaseId(testCreateRequest.getTestCaseId());
@@ -68,9 +83,13 @@ public class TestService {
         return testMapper.toResponse(test);
     }
 
-    public void deleteTestById(Long id, String correlationId) {
+    public void deleteTestById(Long id, String correlationId, AuthInfo authInfo) {
         var test = testRepository.findById(id)
                 .orElseThrow(() -> new NoSuchTestByIdException(id));
+
+        if (hasNoAccessToManageTest(test.getAuthorId(), test.getProjectId(), authInfo)) {
+            throw new AccessToManageEntityDeniedException("Test", "/projects");
+        }
 
         testRepository.delete(test);
 
@@ -82,10 +101,14 @@ public class TestService {
         );
     }
 
-    public TestResponse updateTest(Long id, TestUpdateRequest testUpdateRequest, String correlationId) {
+    public TestResponse updateTest(Long id, TestUpdateRequest testUpdateRequest, String correlationId, AuthInfo authInfo) {
 
         var test = testRepository.findById(id)
                 .orElseThrow(() -> new NoSuchTestByIdException(id));
+
+        if (hasNoAccessToManageTest(test.getAuthorId(), test.getProjectId(), authInfo)) {
+            throw new AccessToManageEntityDeniedException("Test", "/projects");
+        }
 
         var testData = new TestUpdatedEvent.Data();
 
@@ -176,5 +199,11 @@ public class TestService {
                 tests.map(testMapper::toResponse),
                 URI.create("/tests")
         );
+    }
+
+    private boolean hasNoAccessToManageTest(Long ownerId, Long projectId, AuthInfo authInfo) {
+        Long projectOwnerId = projectRepository.findProjectOwnerIdById(projectId);
+
+        return !(isOwner(authInfo, ownerId) || hasAuthority(authInfo, AuthorityName.ADMIN.name()) || isOwner(authInfo, projectOwnerId));
     }
 }
